chore: make checks, rename rules dir
