+++ packages/jhehemann/customs/infer_rules/component.yaml
