+++ packages/napthaai/customs/prediction_request_rag/component.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
   __init__.py: bafybeibt7f7crtwvmkg7spy3jhscmlqltvyblzp32g6gj44v7tlo5lycuq
-  prediction_request_rag.py: bafybeie4rafsx7m7mfkdgd2n2ovfm3gwk5ybjr5cz6s7r3oz545ag2wwuy
+  prediction_request_rag.py: bafybeibx2h5fblewfjqj7pyickfa5gj4dks3batxqyu5bjnjcvusfs2bje
 fingerprint_ignore_patterns: []
 entry_point: prediction_request_rag.py
 callable: run
